Refactor: data-ocr 모듈의 의존성 주입 및 import 오류 수정
- OcrResultEntityMapper 파일에서 detailItems이 OcrDetailItem의 리스트로 바뀌는 것에 따라 코드 수정
