+++ data/ocr/build.gradle.kts
@@ -1,3 +1,6 @@
 android {
     namespace = "com.kappzzang.jeongsan.ocr"
 }
+dependencies {
+    implementation(project(":domain:ocr"))
+}
