+++ data/ocr/src/main/java/com/kappzzang/jeongsan/datasource/ReceiptCaptureFakeDatasource.kt
@@ -1,7 +1,7 @@
 package com.kappzzang.jeongsan.datasource
 
-import com.kappzzang.jeongsan.data.entity.OcrResultEntity
-import com.kappzzang.jeongsan.data.entity.OcrResultItem
+import com.kappzzang.jeongsan.entity.OcrResultEntity
+import com.kappzzang.jeongsan.entity.OcrResultItem
 import javax.inject.Inject
 import kotlinx.coroutines.delay
