+++ data/ocr/src/main/java/com/kappzzang/jeongsan/entity/OcrResultEntity.kt
@@ -1,7 +1,5 @@
 package com.kappzzang.jeongsan.entity
 
-data class OcrResultItem(val name: String, val quantity: Int, val unitPrice: Int)
-
 data class OcrResultEntity(
     val title: String,
     val paymentTime: String,
