+++ data/ocr/src/main/java/com/kappzzang/jeongsan/entity/OcrResultItem.kt
@@ -1,0 +1,3 @@
+package com.kappzzang.jeongsan.entity
+
+data class OcrResultItem(val name: String, val quantity: Int, val unitPrice: Int)
