+++ data/ocr/src/main/java/com/kappzzang/jeongsan/mapper/OcrResultEntityMapper.kt
@@ -1,21 +1,20 @@
 package com.kappzzang.jeongsan.mapper
 
-import com.kappzzang.jeongsan.domain.model.OcrResultResponse
-import com.kappzzang.jeongsan.model.ExpenseDetailItem
+import com.kappzzang.jeongsan.entity.OcrResultEntity
+import com.kappzzang.jeongsan.model.OcrDetailItem
+import com.kappzzang.jeongsan.model.OcrResultResponse
 import java.sql.Timestamp
 
 object OcrResultEntityMapper {
-    fun mapOcrResultEntityToModel(entity: com.kappzzang.jeongsan.entity.OcrResultEntity) =
+    fun mapOcrResultEntityToModel(entity: OcrResultEntity) =
         OcrResultResponse.OcrSuccess(
             name = entity.title,
             paymentTime = Timestamp.valueOf(entity.paymentTime),
-            detailItems = entity.items.mapIndexed { index, item ->
-                com.kappzzang.jeongsan.model.ExpenseDetailItem(
-                    id = index.toString(),
-                    itemQuantity = item.quantity,
+            detailItems = entity.items.map { item ->
+                OcrDetailItem(
                     itemName = item.name,
                     itemPrice = item.unitPrice,
-                    selectedQuantity = 0
+                    itemQuantity = item.quantity
                 )
             }
         )
