+++ data/ocr/src/main/java/com/kappzzang/jeongsan/repositoryimpl/ReceiptCaptureRepositoryImpl.kt
@@ -1,16 +1,18 @@
 package com.kappzzang.jeongsan.repositoryimpl
 
-import com.kappzzang.jeongsan.domain.model.OcrResultResponse
-import com.kappzzang.jeongsan.domain.repository.ReceiptCaptureRepository
+import com.kappzzang.jeongsan.datasource.ReceiptCaptureFakeDatasource
+import com.kappzzang.jeongsan.mapper.OcrResultEntityMapper
+import com.kappzzang.jeongsan.model.OcrResultResponse
+import com.kappzzang.jeongsan.repository.ReceiptCaptureRepository
 import javax.inject.Inject
 
 class ReceiptCaptureRepositoryImpl @Inject constructor(
-    private val dataSource: com.kappzzang.jeongsan.datasource.ReceiptCaptureFakeDatasource
+    private val dataSource: ReceiptCaptureFakeDatasource
 ) : ReceiptCaptureRepository {
 
     override suspend fun getAnalyzedReceiptImage(encodedReceiptImage: String): OcrResultResponse {
         val result = dataSource.sendToOcrServerAndGetResult(encodedReceiptImage)
 
-        return com.kappzzang.jeongsan.mapper.OcrResultEntityMapper.mapOcrResultEntityToModel(result)
+        return OcrResultEntityMapper.mapOcrResultEntityToModel(result)
     }
 }
